Fix timestamp format

+++ src/gameplay_logic/BmsNotes.cpp
@@ -6,17 +6,17 @@
 
 namespace gameplay_logic {
 auto
-BmsNotes::getVisibleNotes() const -> const QList<QList<int>>&
+BmsNotes::getVisibleNotes() const -> const QList<QList<int64_t>>&
 {
     return visibleNotes;
 }
 auto
-BmsNotes::getInvisibleNotes() const -> const QList<QList<int>>&
+BmsNotes::getInvisibleNotes() const -> const QList<QList<int64_t>>&
 {
     return invisibleNotes;
 }
-BmsNotes::BmsNotes(QList<QList<int>> visibleNotes,
-                   QList<QList<int>> invisibleNotes,
+BmsNotes::BmsNotes(QList<QList<int64_t>> visibleNotes,
+                   QList<QList<int64_t>> invisibleNotes,
                    QObject* parent)
   : QObject(parent)
   , visibleNotes(std::move(visibleNotes))
